Reference/openflexure-colour-correction/pyproject.toml
[project]
name = "openflexure_colour_correction"
version = "0.0.1"
authors = [
  { name="Richard Bowman", email="richard.bowman@cantab.net" },
  { name="Julian Stirling", email="julian@julianstirling.co.uk" },
]
description = "Algorithms for colour correcting OpenFlexure images"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Operating System :: OS Independent",
]
dependencies = [
  "numpy>=1.20",
  "scipy ~=1.14",
  "pillow ~= 10.4",
  "piexif",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "pytest-mock",
  "mypy[reports]",
  "mypy-gitlab-code-quality",
  "ruff>=0.1.3",
  "scipy-stubs",
  "lxml",
  "codespell",
]

[project.urls]
"Homepage" = "https://gitlab.com/openflexure/openflexure/openflexure-colour-correction"

[project.scripts]
openflexure-colour-correction = "openflexure_colour_correction.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = [
    "openflexure_colour_correction"
]
artifacts = ["openflexure_colour_correction/data/unmix_tensor.npy"]

[tool.hatch.build.targets.wheel]
artifacts = ["openflexure_colour_correction/data/unmix_tensor.npy"]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.format]
# Use native line endings for all files
line-ending = "native"

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", #pycodestyle warnings
    "F", # PyFlakes
    "PL", # Pylint (a subset of, catches far less than pylint does)
    "B", # Flake8 bugbear
    "A", # Flake8 builtins checker
    "C", # Flake8 comprehensions
    "FIX", #TODO/FIXME's in code. These should be added during development for ongoing tasks.
           # If they are not fixed before merge they should be converted to GitLab issues
    "LOG", # Flake8 logging issues (pedantic logger formatting issues can be added with "G")
    "PT", # pytest linting
    "RET", # Consistent clear return statements
    "RSE", # Raise parentheses
    "SIM", # Simplifications detected
    "ARG", # unused arguments
    "C90", # McCabe complexity!
    "NPY", # Numpy linting
    "N", # PEP8 naming
    "D", # Docstring checks
    "ERA001", # Commented out code!
]

ignore = [
    "E501", # Ignore long lines for now
    "D203",  # incompatible with D204
    "D213",  # incompatible with D212
    "D400",  # A stricter version of #415 that doesn't allow !
    "PLR2004", # Magic numbers in comparisons. Would be fine if we could configure 2
               # and 255 not to be "magic"
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "B018", # Complaining about useless attribute access in tests, but we need them to check errors are raised
    "PLR2004", # Pylint doesn't like magic numbers in comparisons. Ignore for tests.
]


[tool.mypy]
warn_return_any = true
warn_unused_configs = true
files = "openflexure_colour_correction"


[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
    "--cov=openflexure_colour_correction",
    "--cov-report=term",
    "--cov-report=xml:coverage.xml",
    "--cov-report=html:htmlcov",
    "--junitxml=report.xml",
    "--ignore=.git",
    "--ignore=build",
]


pythonpath = ["."]
testpaths = ["tests"]


